Сделал оплату

+++ models/users.py
@@ -22,7 +22,7 @@ class Payments(Base):
 
     id = Column(Integer, primary_key=True)
     username = Column(String, nullable=False)
-    price = Column(Float, nullable=False)
+    price = Column(Integer, nullable=False)
     description = Column(String, nullable=False)
     data = Column(JSON, nullable=False)
     date = Column(Date, nullable=False, default=datetime.utcnow)
